requirements.txt
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
pytz==2023.3.post1
line-bot-sdk==3.9.0
google-cloud-bigquery==3.19.0